refactor: 데이터베이스 관련 객체 패키지 정리
